11.3 commit postman tests pass/failed: 188/137 need more refactor

+++ src/main/java/ru/yandex/practicum/filmorate/controller/MpaRatingController.java
@@ -25,7 +25,7 @@ public class MpaRatingController {
     }
 
     @GetMapping("/{id}")
-    public MpaRating getMpaRatingById(@PathVariable Long id) {
+    public MpaRating getMpaRatingById(@PathVariable Integer id) {
         return mpaRatingService.getMpaRatingById(id);
     }
